Consolidating common code for displaying usage
